+++ src/math/math_utility.h
@@ -9,30 +9,30 @@
 #include "../math/vector_3.h"
 
 namespace MathUtility {
-	constexpr double infinity = std::numeric_limits<double>::infinity();
-	constexpr double pi = 3.1415926535897932385;
+	constexpr float infinity = std::numeric_limits<float>::infinity();
+	constexpr float pi = 3.1415926535897932385f;
 
-	inline double degrees_to_radians(double degrees) {
-		return degrees * pi / 180.0;
+	inline float degrees_to_radians(float degrees) {
+		return degrees * pi / 180.f;
 	}
 
-	inline double random_normalized() {
-		static std::uniform_real_distribution<double> distribution(0.0, 1.0);
+	inline float random_normalized() {
+		static std::uniform_real_distribution<float> distribution(0.f, 1.f);
 		static std::mt19937 generator;
 		return distribution(generator);
 	}
 
-	inline double random_range(double min, double max) {
+	inline float random_range(float min, float max) {
 		return min + (max - min) * random_normalized();
 	}
 
 	inline int random_range(int min, int max) {
-		return int(random_range(double(min), double(max)));
+		return int(random_range(static_cast<float>(min), static_cast<float>(max)));
 	}
 
 	inline Vector3 random_in_unit_sphere() {
 		while(true) {
-			auto p = Vector3::random(-1.0, 1.0);
+			auto p = Vector3::random(-1.f, 1.f);
 			if(p.squared_magnitude() < 1) return p;
 		}
 	}
@@ -44,18 +44,18 @@ namespace MathUtility {
 	inline Vector3 random_on_hemisphere(const Vector3& normal) {
 		Vector3 on_unit_sphere = random_unit_vector();
 
-		if(on_unit_sphere.dot(normal) > 0.0) return on_unit_sphere;
+		if(on_unit_sphere.dot(normal) > 0.f) return on_unit_sphere;
 		return -on_unit_sphere;
 	}
 
 	inline Vector3 random_in_unit_disk() {
 		while(true) {
-			auto p = Vector3{random_range(-1.0,1.0), random_range(-1.0, 1.0), 0.0};
-			if(p.squared_magnitude() < 1.0) return p;
+			auto p = Vector3{random_range(-1.f,1.f), random_range(-1.f, 1.f), 0.f};
+			if(p.squared_magnitude() < 1.f) return p;
 		}
 	}
 
-	inline double linear_to_gamma(double linear) {
+	inline float linear_to_gamma(float linear) {
 		return sqrt(linear);
 	}
 
@@ -68,7 +68,7 @@ namespace MathUtility {
 		g = linear_to_gamma(g);
 		b = linear_to_gamma(b);
 
-		Interval intensity{0.000, 0.999};
+		Interval intensity{0.f, 0.999f};
 		constexpr int max_value = 256;
 
 		out << static_cast<int>(max_value * intensity.clamped(r)) << ' '
