+++ src/math/vector_3.h
@@ -5,26 +5,26 @@
 struct Vector3 {
 	union {
 		struct {
-			double x;
-			double y;
-			double z;
+			float x;
+			float y;
+			float z;
 		};
-		double components[3] = {0.f};
+		float components[3] = {0.f};
 	};
 
 	Vector3();
-	explicit Vector3(double value);
-	Vector3(double x, double y, double z);
+	explicit Vector3(float value);
+	Vector3(float x, float y, float z);
 
 	static Vector3 random();
-	static Vector3 random(double min, double max);
+	static Vector3 random(float min, float max);
 
-	double& operator[](int i);
-	const double& operator[](int i) const;
+	float& operator[](int i);
+	const float& operator[](int i) const;
 
-	Vector3& operator*=(double s);
+	Vector3& operator*=(float s);
 
-	Vector3& operator/=(double s);
+	Vector3& operator/=(float s);
 
 	Vector3& operator+=(const Vector3& v);
 
@@ -34,27 +34,27 @@ struct Vector3 {
 	[[nodiscard]] Vector3 normalized() const;
 	[[nodiscard]] bool is_near_zero() const;
 
-	void clamp(double clamped_magnitude);
-	[[nodiscard]] Vector3 clamped(double clamped_magnitude) const;
+	void clamp(float clamped_magnitude);
+	[[nodiscard]] Vector3 clamped(float clamped_magnitude) const;
 
-	[[nodiscard]] double squared_magnitude() const;
-	[[nodiscard]] double magnitude() const;
+	[[nodiscard]] float squared_magnitude() const;
+	[[nodiscard]] float magnitude() const;
 
 	[[nodiscard]] Vector3 cross(const Vector3& v) const;
 	[[nodiscard]] Vector3 reflect(const Vector3& n) const;
-	[[nodiscard]] Vector3 refract(const Vector3& n, double etai_over_etat);
+	[[nodiscard]] Vector3 refract(const Vector3& n, float etai_over_etat);
 
-	[[nodiscard]] double dot(const Vector3& v) const;
+	[[nodiscard]] float dot(const Vector3& v) const;
 };
 
 using Point3 = Vector3;
 using Color = Vector3;
 
 Vector3 operator*(Vector3 a, const Vector3& b);
-Vector3 operator*(Vector3 v, double s);
-Vector3 operator*(double s, Vector3 v);
+Vector3 operator*(Vector3 v, float s);
+Vector3 operator*(float s, Vector3 v);
 
-Vector3 operator/(Vector3 v, double s);
+Vector3 operator/(Vector3 v, float s);
 
 Vector3 operator+(Vector3 a, const Vector3& b);
