+++ src/scenes/five_spheres.h
@@ -12,7 +12,7 @@
 #include "../math/vector_3.h"
 
 namespace {
-	static void create_sphere(HittableList& scene, const Point3& position, double r, std::shared_ptr<Material> material) {
+	static void create_sphere(HittableList& scene, const Point3& position, float r, std::shared_ptr<Material> material) {
 		scene.add(std::make_shared<Sphere>(position, r, material));
 	}
 }
@@ -21,31 +21,31 @@ static HittableList five_spheres() {
 	HittableList scene;
 
 	// Ground
-	auto ground_material = std::make_shared<Lambertian>(Color{0.13, 0.69, 0.3});
-	create_sphere(scene, {0.0, -1002.0, 0.0}, 1000.0, ground_material);
+	auto ground_material = std::make_shared<Lambertian>(Color{0.13f, 0.69f, 0.3f});
+	create_sphere(scene, {0.f, -1002.f, 0.f}, 1000.f, ground_material);
 
 	// 5 Spheres placed in a trapezoid shape
-	auto distance = 1.0;
-	auto radius = 0.8;
-	auto y_offset = -0.25;
+	auto distance = 1.f;
+	auto radius = 0.8f;
+	auto y_offset = -0.25f;
 
-	auto tri_x = cos(MathUtility::pi / 6.0) * distance;
-	auto tri_y = sin(MathUtility::pi / 6.0) * distance;
+	auto tri_x = cos(MathUtility::pi / 6.f) * distance;
+	auto tri_y = sin(MathUtility::pi / 6.f) * distance;
 
-	auto gold_metal_material = std::make_shared<Metal>(Color{1.0, 0.78, 0.08}, 0.05);
-	create_sphere(scene, {-tri_x * 2.0, distance + y_offset, 0.0}, radius, gold_metal_material);
+	auto gold_metal_material = std::make_shared<Metal>(Color{1.f, 0.78f, 0.08f}, 0.05f);
+	create_sphere(scene, {-tri_x * 2.f, distance + y_offset, 0.f}, radius, gold_metal_material);
 
-	auto red_diffuse_material = std::make_shared<Lambertian>(Color{0.9, 0.1, 0.1});
-	create_sphere(scene, {-tri_x, -tri_y + y_offset, 0.0}, radius, red_diffuse_material);
+	auto red_diffuse_material = std::make_shared<Lambertian>(Color{0.9f, 0.1f, 0.1f});
+	create_sphere(scene, {-tri_x, -tri_y + y_offset, 0.f}, radius, red_diffuse_material);
 
-	auto green_diffuse_material = std::make_shared<Lambertian>(Color{0.71, 0.9, 0.11});
-	create_sphere(scene, {0.0, distance + y_offset, 0.0}, radius, green_diffuse_material);
+	auto green_diffuse_material = std::make_shared<Lambertian>(Color{0.71f, 0.9f, 0.11f});
+	create_sphere(scene, {0.f, distance + y_offset, 0.f}, radius, green_diffuse_material);
 
-	auto blue_diffuse_material = std::make_shared<Lambertian>(Color{0.0, 0.6, 0.9});
-	create_sphere(scene, {tri_x, -tri_y + y_offset, 0.0}, radius, blue_diffuse_material);
+	auto blue_diffuse_material = std::make_shared<Lambertian>(Color{0.f, 0.6f, 0.9f});
+	create_sphere(scene, {tri_x, -tri_y + y_offset, 0.f}, radius, blue_diffuse_material);
 
-	auto silver_metal_material = std::make_shared<Metal>(Color{0.9, 0.9, 0.9}, 0.3);
-	create_sphere(scene, {tri_x * 2.0, distance + y_offset, 0.0}, radius, silver_metal_material);
+	auto silver_metal_material = std::make_shared<Metal>(Color{0.9f, 0.9f, 0.9f}, 0.3f);
+	create_sphere(scene, {tri_x * 2.f, distance + y_offset, 0.f}, radius, silver_metal_material);
 
 	return scene;
 }
